Move LakeDataSource and LakeDataSourceProvider to cpp file
