+++ be/src/connector/lake_connector.cpp
@@ -4,8 +4,13 @@
 
 #include "common/constexpr.h"
 #include "exec/vectorized/connector_scan_node.h"
+#include "exec/vectorized/olap_scan_prepare.h"
+#include "gen_cpp/InternalService_types.h"
+#include "runtime/global_dict/parser.h"
 #include "storage/column_predicate_rewriter.h"
+#include "storage/conjunctive_predicates.h"
 #include "storage/lake/tablet_manager.h"
+#include "storage/lake/tablet_reader.h"
 #include "storage/predicate_parser.h"
 #include "storage/projection_iterator.h"
 #include "util/starrocks_metrics.h"
@@ -15,19 +20,132 @@ using namespace vectorized;
 
 // ================================
 
-DataSourceProviderPtr LakeConnector::create_data_source_provider(vectorized::ConnectorScanNode* scan_node,
-                                                                 const TPlanNode& plan_node) const {
-    return std::make_unique<LakeDataSourceProvider>(scan_node, plan_node);
-}
+class LakeDataSourceProvider;
+
+// TODO: support parallel scan within a single tablet
+class LakeDataSource final : public DataSource {
+public:
+    explicit LakeDataSource(const LakeDataSourceProvider* provider, const TScanRange& scan_range);
+    ~LakeDataSource() override;
+
+    Status open(RuntimeState* state) override;
+    void close(RuntimeState* state) override;
+    Status get_next(RuntimeState* state, vectorized::ChunkPtr* chunk) override;
+
+    int64_t raw_rows_read() const override { return _raw_rows_read; }
+    int64_t num_rows_read() const override { return _num_rows_read; }
+    int64_t num_bytes_read() const override { return _bytes_read; }
+
+    // TODO:
+    // Return last bytes of Scan or Exchange Data, then reset it
+    // int64_t last_spent_cpu_time_ns() override;
+
+private:
+    Status get_tablet(const TInternalScanRange& scan_range);
+    Status init_global_dicts(vectorized::TabletReaderParams* params);
+    Status init_unused_output_columns(const std::vector<std::string>& unused_output_columns);
+    Status init_scanner_columns(std::vector<uint32_t>& scanner_columns);
+    void decide_chunk_size();
+    Status init_reader_params(const std::vector<OlapScanRange*>& key_ranges,
+                              const std::vector<uint32_t>& scanner_columns, std::vector<uint32_t>& reader_columns);
+    Status init_tablet_reader(RuntimeState* state);
+    Status build_scan_range(RuntimeState* state);
+    void init_counter(RuntimeState* state);
+    void update_realtime_counter(vectorized::Chunk* chunk);
+    void update_counter();
+
+private:
+    const LakeDataSourceProvider* _provider;
+    const TInternalScanRange _scan_range;
+
+    Status _status = Status::OK();
+    // The conjuncts couldn't push down to storage engine
+    std::vector<ExprContext*> _not_push_down_conjuncts;
+    vectorized::ConjunctivePredicates _not_push_down_predicates;
+    std::vector<uint8_t> _selection;
+
+    ObjectPool _obj_pool;
+
+    RuntimeState* _runtime_state = nullptr;
+    const std::vector<SlotDescriptor*>* _slots = nullptr;
+    std::vector<std::unique_ptr<OlapScanRange>> _key_ranges;
+    std::vector<OlapScanRange*> _scanner_ranges;
+    vectorized::OlapScanConjunctsManager _conjuncts_manager;
+    vectorized::DictOptimizeParser _dict_optimize_parser;
+
+    std::shared_ptr<const TabletSchema> _tablet_schema;
+    int64_t _version = 0;
+    vectorized::TabletReaderParams _params{};
+    std::shared_ptr<lake::TabletReader> _reader;
+    // projection iterator, doing the job of choosing |_scanner_columns| from |_reader_columns|.
+    std::shared_ptr<vectorized::ChunkIterator> _prj_iter;
+
+    const std::vector<std::string>* _unused_output_columns = nullptr;
+    std::unordered_set<uint32_t> _unused_output_column_ids;
+    // For release memory.
+    using PredicatePtr = std::unique_ptr<vectorized::ColumnPredicate>;
+    std::vector<PredicatePtr> _predicate_free_pool;
+
+    // slot descriptors for each one of |output_columns|.
+    std::vector<SlotDescriptor*> _query_slots;
+
+    // The following are profile meatures
+    int64_t _num_rows_read = 0;
+    int64_t _raw_rows_read = 0;
+    int64_t _bytes_read = 0;
+
+    RuntimeProfile::Counter* _bytes_read_counter = nullptr;
+    RuntimeProfile::Counter* _rows_read_counter = nullptr;
+
+    RuntimeProfile::Counter* _expr_filter_timer = nullptr;
+    RuntimeProfile::Counter* _scan_timer = nullptr;
+    RuntimeProfile::Counter* _create_seg_iter_timer = nullptr;
+    RuntimeProfile::Counter* _tablet_counter = nullptr;
+    RuntimeProfile::Counter* _reader_init_timer = nullptr;
+    RuntimeProfile::Counter* _io_timer = nullptr;
+    RuntimeProfile::Counter* _read_compressed_counter = nullptr;
+    RuntimeProfile::Counter* _decompress_timer = nullptr;
+    RuntimeProfile::Counter* _read_uncompressed_counter = nullptr;
+    RuntimeProfile::Counter* _raw_rows_counter = nullptr;
+    RuntimeProfile::Counter* _pred_filter_counter = nullptr;
+    RuntimeProfile::Counter* _del_vec_filter_counter = nullptr;
+    RuntimeProfile::Counter* _pred_filter_timer = nullptr;
+    RuntimeProfile::Counter* _chunk_copy_timer = nullptr;
+    RuntimeProfile::Counter* _seg_init_timer = nullptr;
+    RuntimeProfile::Counter* _zm_filtered_counter = nullptr;
+    RuntimeProfile::Counter* _bf_filtered_counter = nullptr;
+    RuntimeProfile::Counter* _seg_zm_filtered_counter = nullptr;
+    RuntimeProfile::Counter* _sk_filtered_counter = nullptr;
+    RuntimeProfile::Counter* _block_seek_timer = nullptr;
+    RuntimeProfile::Counter* _block_seek_counter = nullptr;
+    RuntimeProfile::Counter* _block_load_timer = nullptr;
+    RuntimeProfile::Counter* _block_load_counter = nullptr;
+    RuntimeProfile::Counter* _block_fetch_timer = nullptr;
+    RuntimeProfile::Counter* _index_load_timer = nullptr;
+    RuntimeProfile::Counter* _read_pages_num_counter = nullptr;
+    RuntimeProfile::Counter* _cached_pages_num_counter = nullptr;
+    RuntimeProfile::Counter* _bi_filtered_counter = nullptr;
+    RuntimeProfile::Counter* _bi_filter_timer = nullptr;
+    RuntimeProfile::Counter* _pushdown_predicates_counter = nullptr;
+    RuntimeProfile::Counter* _rowsets_read_count = nullptr;
+    RuntimeProfile::Counter* _segments_read_count = nullptr;
+    RuntimeProfile::Counter* _total_columns_data_page_count = nullptr;
+};
 
 // ================================
 
-LakeDataSourceProvider::LakeDataSourceProvider(vectorized::ConnectorScanNode* scan_node, const TPlanNode& plan_node)
-        : _scan_node(scan_node), _t_olap_scan_node(plan_node.olap_scan_node) {}
+class LakeDataSourceProvider final : public DataSourceProvider {
+public:
+    friend class LakeDataSource;
+    LakeDataSourceProvider(vectorized::ConnectorScanNode* scan_node, const TPlanNode& plan_node);
+    ~LakeDataSourceProvider() override = default;
 
-DataSourcePtr LakeDataSourceProvider::create_data_source(const TScanRange& scan_range) {
-    return std::make_unique<LakeDataSource>(this, scan_range);
-}
+    DataSourcePtr create_data_source(const TScanRange& scan_range) override;
+
+protected:
+    vectorized::ConnectorScanNode* _scan_node;
+    const TOlapScanNode _t_olap_scan_node;
+};
 
 // ================================
 
@@ -459,6 +577,22 @@ void LakeDataSource::update_counter() {
         COUNTER_UPDATE(c1, _reader->stats().del_filter_ns);
         COUNTER_UPDATE(c2, _reader->stats().rows_del_filtered);
     }
+}
+
+// ================================
+
+LakeDataSourceProvider::LakeDataSourceProvider(vectorized::ConnectorScanNode* scan_node, const TPlanNode& plan_node)
+        : _scan_node(scan_node), _t_olap_scan_node(plan_node.olap_scan_node) {}
+
+DataSourcePtr LakeDataSourceProvider::create_data_source(const TScanRange& scan_range) {
+    return std::make_unique<LakeDataSource>(this, scan_range);
+}
+
+// ================================
+
+DataSourceProviderPtr LakeConnector::create_data_source_provider(vectorized::ConnectorScanNode* scan_node,
+                                                                 const TPlanNode& plan_node) const {
+    return std::make_unique<LakeDataSourceProvider>(scan_node, plan_node);
 }
 
 } // namespace starrocks::connector
